chore: app.js -> app.tsx
